project_agora/requirements.txt
# Production dependencies for docker build
google-adk>=1.0.0
google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0
google-cloud-storage>=2.10.0
pydantic>=2.11.3
python-dotenv>=1.0.1
sqlglot>=26.10.1
db-dtypes>=1.4.2
requests>=2.31.0
llama-index>=0.12
tabulate>=0.9.0

# Development dependencies
pytest>=8.3.5
pytest-asyncio>=0.26.0 
markdownify
playwright